[PW-6532] Declarative schema (#1505)
* [PW-6532] Declarative schema

* [PW-6532] Declarative data patch for 8.0.0

Dropping patch for v2.4.4 since that version is not officially supported.

* [PW-6532] Declarative data patch for 8.2.1

* [PW-6532] Generated schema allow list

* [PW-6532] Fixing notification schema

Generated fields had wrong type and/or default values missing.
