+++ Setup/Patch/Data/SplitPaymentsBecomePartialPayments.php
@@ -1,0 +1,94 @@
+<?php
+/**
+ * Adyen Payment module (https://www.adyen.com/)
+ *
+ * Copyright (c) 2022 Adyen BV (https://www.adyen.com/)
+ * See LICENSE.txt for license details.
+ *
+ * Author: Adyen <magento@adyen.com>
+ */
+declare(strict_types=1);
+
+namespace Adyen\Payment\Setup\Patch\Data;
+
+use Magento\Framework\Setup\ModuleDataSetupInterface;
+use Magento\Framework\Setup\Patch\DataPatchInterface;
+use Magento\Framework\Setup\Patch\PatchVersionInterface;
+
+class SplitPaymentsBecomePartialPayments implements DataPatchInterface, PatchVersionInterface
+{
+    /**
+     * @var ModuleDataSetupInterface $moduleDataSetup
+     */
+    private $moduleDataSetup;
+
+    /**
+     * @param ModuleDataSetupInterface $moduleDataSetup
+     */
+    public function __construct(ModuleDataSetupInterface $moduleDataSetup)
+    {
+        $this->moduleDataSetup = $moduleDataSetup;
+    }
+
+    /**
+     * Do Upgrade
+     *
+     * @return void
+     */
+    public function apply()
+    {
+        $this->moduleDataSetup->getConnection()->startSetup();
+        $this->updateSchemaVersion800($this->moduleDataSetup);
+        $this->moduleDataSetup->getConnection()->endSetup();
+    }
+
+    /**
+     * Ensure that new path does not exist before updating
+     *
+     * @param ModuleDataSetupInterface $setup
+     */
+    public function updateSchemaVersion800(ModuleDataSetupInterface $setup)
+    {
+        $partialPaymentsPath = 'payment/adyen_abstract/partial_payments_refund_strategy';
+        $configDataTable = $setup->getTable('core_config_data');
+        $connection = $setup->getConnection();
+
+        $select = $connection->select()
+            ->from($configDataTable)
+            ->where(
+                'path = ?',
+                $partialPaymentsPath
+            );
+
+        $partialPaymentConfig = $connection->fetchRow($select);
+
+        if (is_null($partialPaymentConfig)) {
+            $connection->update(
+                $configDataTable,
+                ['path' => 'payment/adyen_abstract/partial_payments_refund_strategy'],
+                ['path = ?' => 'payment/adyen_abstract/split_payments_refund_strategy']
+            );
+        }
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function getAliases()
+    {
+        return [];
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public static function getDependencies()
+    {
+        return [];
+    }
+
+    public static function getVersion()
+    {
+        return '8.0.0';
+    }
+}
